Random number generation simplified
