+++ src/Makefile.in
@@ -37,15 +37,13 @@ NVCC       	   = @NVCC@
 srcdir	 = @srcdir@
 VPATH	  = @srcdir@
 
-OBJS=main.o phase_loss_opt.o io.o problem.o mpi_optalg.o mpi_pso.o mpi_de.o
+OBJS=main.o phase_loss_opt.o io.o problem.o mpi_optalg.o mpi_pso.o mpi_de.o rng.o
 
 ifdef CUDA_LIBS
 	OBJS+=rng_gpu.cu.co
 else 
   ifdef VSL_LIBS
 		OBJS+=rng_vsl.o
-	else
-		OBJS+=rng_simple.o
 	endif
 endif
