+++ src/mpi_optalg.h
@@ -58,4 +58,44 @@ class OptAlg {
 
         bool goal;
     };
+
+class DE : public OptAlg {
+    public:
+        DE() {};
+        DE(Problem *problem_ptr): F(0.1), Cr(0.6) {
+            this->prob = problem_ptr;
+            this->num = this->prob->num;
+            }
+        ~DE() {};
+
+        void put_to_best(int my_rank, int total_pop, int nb_proc);
+        void combination(int my_rank, int total_pop, int nb_proc);
+        void selection(int my_rank, int total_pop, int nb_proc);
+        void write_param(double *param_array);
+        void fit_to_global();
+        void find_global(int my_rank, int total_pop, int nb_proc) {};
+
+    private:
+        double F, Cr;
+    };
+    
+class PSO : public OptAlg {
+    public:
+        PSO() {};
+        PSO(Problem *problem_ptr): w(0.8), phi1(0.6), phi2(1.0), v_max(0.2) {
+            this->prob = problem_ptr;
+            this->num = this->prob->num;
+            }
+        ~PSO() {};
+
+        void put_to_best(int my_rank, int total_pop, int nb_proc);
+        void combination(int my_rank, int total_pop, int nb_proc);
+        void selection(int my_rank, int total_pop, int nb_proc);
+        void write_param(double *param_array);
+        void fit_to_global() {};
+        void find_global(int my_rank, int total_pop, int nb_proc);
+
+    private:
+        double w, phi1, phi2, v_max;
+    };    
 #endif // OPTALG_H
